more refactoring to clean up Trace.ts
